Feat: Adding donult chart for results

+++ database/migrations/2024_04_15_234350_create_dados_sensores_table.php
@@ -14,11 +14,11 @@ class CreateDadosSensoresTable extends Migration
             $table->decimal('soilTemperature', 8, 2)->nullable();
             $table->decimal('airHumidity', 8, 2)->nullable();
             $table->decimal('airTemperature', 8, 2)->nullable();
-            $table->unsignedInteger('soilConductivity')->nullable();
+            $table->decimal('soilConductivity')->nullable();
             $table->decimal('soilPH', 8, 2)->nullable();
-            $table->unsignedInteger('nitrogen')->nullable();
-            $table->unsignedInteger('phosphorus')->nullable();
-            $table->unsignedInteger('potassium')->nullable();
+            $table->decimal('nitrogen')->nullable();
+            $table->decimal('phosphorus')->nullable();
+            $table->decimal('potassium')->nullable();
             $table->timestamps();
         });
     }
